thinking on designing how to get a local address for a service.
